Day 9 - Update Intcode computer to operate on BigIntegers, add ~64k BigIntegers worth of memory, and support for relative pointers

+++ intcode-computer/src/main/kotlin/intcode/ops/Add.kt
@@ -1,13 +1,14 @@
 package intcode.ops
 
 import intcode.Param
+import java.math.BigInteger
 
 class Add(private val a: Param,
           private val b: Param,
           private val result: Param) : Operation {
-    override fun eval(): (Int) -> Int {
+    override fun eval(): (BigInteger) -> BigInteger {
         result.write(a.read() + b.read())
-        return { it + 4 }
+        return { it + 4.toBigInteger() }
     }
 
     override fun toString(): String {
