+++ intcode-computer/src/main/kotlin/intcode/ops/InputToMemory.kt
@@ -1,11 +1,12 @@
 package intcode.ops
 
 import intcode.Param
+import java.math.BigInteger
 
-class InputToMemory(private val param: Param, private val input: () -> Int) : Operation {
-    override fun eval(): (Int) -> Int {
+class InputToMemory(private val param: Param, private val input: () -> BigInteger) : Operation {
+    override fun eval(): (BigInteger) -> BigInteger {
         param.write(input())
-        return { it + 2 }
+        return { it + 2.toBigInteger() }
     }
 
     override fun toString(): String {
