+++ intcode-computer/src/main/kotlin/intcode/ops/MemoryToOutput.kt
@@ -1,11 +1,12 @@
 package intcode.ops
 
 import intcode.Param
+import java.math.BigInteger
 
-class MemoryToOutput(private val param: Param, private val output: (Int) -> Unit) : Operation {
-    override fun eval(): (Int) -> Int {
+class MemoryToOutput(private val param: Param, private val output: (BigInteger) -> Unit) : Operation {
+    override fun eval(): (BigInteger) -> BigInteger {
         output(param.read())
-        return { it + 2 }
+        return { it + 2.toBigInteger() }
     }
 
     override fun toString(): String {
